server/CustomAPI/requirements.txt
responder==2.0.7
starlette==0.13
httptools==0.1.1
tweepy
pandas
python-dotenv
requests
numpy
opencv-python
git+https://github.com/JustAnotherArchivist/snscrape.git